requirements.txt
fastapi
uvicorn
docling
pydantic
pydantic-settings
python-multipart
PyJWT
python-jose[cryptography]
supabase
resend
pinecone
python-dotenv
langgraph
langchain-core
langchain-groq
langchain
pyyaml
email-validator
